SW-6603 Fix styling of form
